Add mapping table
A new mapping table that can be modified on runtime allows to change the actions
that are performed if a key event is read from the Wii Remote.
This also adds a "motion" field which specifies whether the Wii Remote emulates
a relative or absolute motion input or no motion input at all.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>