Add support for hotplug events
We now watch for device hotplug events. This allows plugging extensions
during runtime and we correctly initialize them.

We do this by re-opening all required interfaces on every hotplug event.
The xwiimote library handles this correctly internally, so we don't end
up closing and reopening the interfaces all the time.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>