Fix size calculations in xwiimote_rm_dev()
We move elements in an array of pointers, not array of device objects.
Hence, we need to use sizeof(*iter) instead of sizeof(*dev).

This fixes a serious out-of-bounds access during device unplugging.

Also increase the buffer size by one so the last entry is guaranteed to be
NULL. Otherwise, if MAXDEVICES are used, we will have no NULL termination.
This also eliminates the need to clear the last entry during
xwiimote_rm_dev().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>